+++ lib/koji_helpers/klean/tool.py
@@ -23,6 +23,7 @@ from logging import getLogger
 from koji_helpers import CONFIG
 from koji_helpers.config import Configuration
 from koji_helpers.klean.distrepo import DistRepoCleaner
+from koji_helpers.klean.scratchbuild import ScratchBuildCleaner
 
 __author__ = """John Florian <jflorian@doubledog.org>"""
 __copyright__ = """2019 John Florian"""
@@ -62,4 +63,5 @@ class KleanTool(object):
         _log.info('started')
         # Using smashd's repo names as dist tags here.
         DistRepoCleaner(self.config.repos, self.config)
+        ScratchBuildCleaner(self.config)
         _log.info('finished')